[FIX] disabled RazerS3 for VS8 and below
git-svn-id: http://svn.seqan.de/seqan/trunk@12600 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/extras/apps/razers3/razers_match_filter.h
@@ -166,7 +166,7 @@ _newLimitDistRange(TMatchFilter const & filter, unsigned histogramId)
         if (max <= 0)
         {
             // do we have enough matches in [0..scoreDistanceRange) ?
-            if (filter.options.purgeAmbiguous && it - itBeg < filter.options.scoreDistanceRange)
+            if (filter.options.purgeAmbiguous && it - itBeg < (int)filter.options.scoreDistanceRange)
                 return CAN_BE_PURGED;
             else
                 return it - itBeg;
